Contract test changes
