+++ pacts/et_sya_api_service-idam_user_details.json
@@ -1,0 +1,119 @@
+{
+  "provider": {
+    "name": "idam_user_details"
+  },
+  "consumer": {
+    "name": "et_sya_api_service"
+  },
+  "interactions": [
+    {
+      "description": "Provider receives a token request and send user details to the API",
+      "request": {
+        "method": "GET",
+        "path": "/details"
+      },
+      "response": {
+        "status": 200,
+        "headers": {
+          "Authorization": "Bearer UserAuthToken",
+          "Content-Type": "application/json; charset=UTF-8"
+        },
+        "body": {
+          "forename": "Joe",
+          "surname": "Bloggs",
+          "roles": [
+            "caseworker",
+            "citizen"
+          ],
+          "active": true,
+          "id": "123432",
+          "email": "joe.bloggs@hmcts.net"
+        },
+        "matchingRules": {
+          "body": {
+            "$.id": {
+              "matchers": [
+                {
+                  "match": "type"
+                }
+              ],
+              "combine": "AND"
+            },
+            "$.forename": {
+              "matchers": [
+                {
+                  "match": "type"
+                }
+              ],
+              "combine": "AND"
+            },
+            "$.surname": {
+              "matchers": [
+                {
+                  "match": "type"
+                }
+              ],
+              "combine": "AND"
+            },
+            "$.email": {
+              "matchers": [
+                {
+                  "match": "type"
+                }
+              ],
+              "combine": "AND"
+            },
+            "$.active": {
+              "matchers": [
+                {
+                  "match": "type"
+                }
+              ],
+              "combine": "AND"
+            },
+            "$.roles[0]": {
+              "matchers": [
+                {
+                  "match": "type"
+                }
+              ],
+              "combine": "AND"
+            },
+            "$.roles[1]": {
+              "matchers": [
+                {
+                  "match": "type"
+                }
+              ],
+              "combine": "AND"
+            }
+          },
+          "header": {
+            "Content-Type": {
+              "matchers": [
+                {
+                  "match": "regex",
+                  "regex": "application/json(;\\s?charset=[\\w\\-]+)?"
+                }
+              ],
+              "combine": "AND"
+            }
+          }
+        }
+      },
+      "providerStates": [
+        {
+          "name": "a user exists"
+        }
+      ]
+    }
+  ],
+  "metadata": {
+    "pactSpecification": {
+      "version": "3.0.0"
+    },
+    "pact-jvm": {
+      "version": "4.1.41"
+    }
+  }
+}
