+++ src/contractTest/java/uk/gov/hmcts/reform/et/syaapi/consumer/test/ccd/StartEventForCitizenConsumerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.et.syaapi.consumer.ccd;
+package uk.gov.hmcts.reform.et.syaapi.consumer.test.ccd;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
@@ -9,21 +9,22 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
-import uk.gov.hmcts.reform.et.syaapi.consumer.SpringBootContractTestBase;
+import uk.gov.hmcts.reform.et.syaapi.consumer.test.SpringBootContractBaseTest;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.springframework.http.HttpMethod.GET;
 
-class StartForCitizenConsumerTest extends SpringBootContractTestBase {
+
+class StartEventForCitizenConsumerTest extends SpringBootContractBaseTest {
 
     @Pact(provider = "ccd_data_store_api_cases", consumer = "et_sya_api_service")
-        RequestResponsePact startForCitizen(PactDslWithProvider builder) {
+    RequestResponsePact startEventForCitizen(PactDslWithProvider builder) {
 
         return  builder
-                .given("A Start case for a Citizen is requested", addCaseTypeJurdisticaton())
-                .uponReceiving("A Start case for a Citizen by calling CCD API")
+                .given("A Start event for a Citizen is requested", addCaseTypeJurisdiction())
+                .uponReceiving("A Start event for a Citizen against CCD API")
                 .path(buildPath())
                 .method(GET.toString())
                 .headers(RESPONSE_HEADERS)
@@ -31,33 +32,33 @@ class StartForCitizenConsumerTest extends SpringBootContractTestBase {
                 .willRespondWith()
                 .matchHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
                 .status(200)
-                .body(buildStartEventResponseWithEmptyCaseDetails(INITIATE_CASE_DRAFT))
+                .body(buildStartEventResponseWithEmptyCaseDetails(UPDATE_CASE_DRAFT))
                 .toPact();
     }
 
     @Test
     @SneakyThrows
-    @PactTestFor(pactMethod = "startForCitizen")
-    void verifyStartForCitizen() {
+    @PactTestFor(pactMethod = "startEventForCitizen")
+    void verifyStartEventForCitizen() {
 
-        StartEventResponse startEventResponse = coreCaseDataApi.startForCitizen(
-            SERVICE_AUTH_TOKEN, AUTH_TOKEN, USER_ID, JURISDICTION_ID, CASE_TYPE_ID, INITIATE_CASE_DRAFT);
+        StartEventResponse startEventResponse = coreCaseDataApi.startEventForCitizen(SERVICE_AUTH_TOKEN, AUTH_TOKEN,
+              USER_ID, JURISDICTION_ID, CASE_TYPE_ID, CASE_ID.toString(), UPDATE_CASE_DRAFT);
 
-        assertThat(startEventResponse.getEventId(), equalTo(INITIATE_CASE_DRAFT));
+        assertThat(startEventResponse.getEventId(), equalTo(UPDATE_CASE_DRAFT));
         assertNotNull(startEventResponse.getCaseDetails());
     }
 
     private String buildPath() {
-        return new StringBuilder()
-            .append("/citizens/")
-            .append(USER_ID)
-            .append("/jurisdictions/")
-            .append(JURISDICTION_ID)
-            .append("/case-types/")
-            .append(CASE_TYPE_ID)
-            .append("/event-triggers/")
-            .append(INITIATE_CASE_DRAFT)
-            .append("/token")
-            .toString();
+        return "/citizens/"
+            + USER_ID
+            + "/jurisdictions/"
+            + JURISDICTION_ID
+            + "/case-types/"
+            + CASE_TYPE_ID
+            + "/cases/"
+            + CASE_ID
+            + "/event-triggers/"
+            + UPDATE_CASE_DRAFT
+            + "/token";
     }
 }
