+++ src/contractTest/java/uk/gov/hmcts/reform/et/syaapi/consumer/test/ccd/SubmitEventForCitizenConsumerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.et.syaapi.consumer.ccd;
+package uk.gov.hmcts.reform.et.syaapi.consumer.test.ccd;
 
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
@@ -12,19 +12,19 @@ import uk.gov.hmcts.et.common.model.ccd.Et1CaseData;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.et.syaapi.consumer.SpringBootContractTestBase;
+import uk.gov.hmcts.reform.et.syaapi.consumer.test.SpringBootContractBaseTest;
 import uk.gov.hmcts.reform.et.syaapi.service.utils.ResourceLoader;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.springframework.http.HttpMethod.POST;
 
-class SubmitEventForCitizenConsumerTest extends SpringBootContractTestBase {
+class SubmitEventForCitizenConsumerTest extends SpringBootContractBaseTest {
 
     @Pact(provider = "ccd_data_store_api_cases", consumer = "et_sya_api_service")
     RequestResponsePact submitEventForCitizen(PactDslWithProvider builder) {
 
         return builder
-            .given("A Submit event for a Citizen is requested", addCaseTypeJurdisticaton())
+            .given("A Submit event for a Citizen is requested", addCaseTypeJurisdiction())
             .uponReceiving("A Submit event for a Citizen against CCD API")
             .path(buildPath())
             .query("ignore-warning=true")
@@ -41,7 +41,7 @@ class SubmitEventForCitizenConsumerTest extends SpringBootContractTestBase {
     @Test
     @SneakyThrows
     @PactTestFor(pactMethod = "submitEventForCitizen")
-    void verifysubmitEventForCitizen() {
+    void verifySubmitEventForCitizen() {
         Et1CaseData caseData = ResourceLoader.fromString("requests/caseData.json", Et1CaseData.class);
 
         CaseDataContent caseDataContent = CaseDataContent.builder()
@@ -59,16 +59,14 @@ class SubmitEventForCitizenConsumerTest extends SpringBootContractTestBase {
     }
 
     private String buildPath() {
-        return new StringBuilder()
-            .append("/citizens/")
-            .append(USER_ID)
-            .append("/jurisdictions/")
-            .append(JURISDICTION_ID)
-            .append("/case-types/")
-            .append(CASE_TYPE_ID)
-            .append("/cases/")
-            .append(CASE_ID)
-            .append("/events")
-            .toString();
+        return "/citizens/"
+            + USER_ID
+            + "/jurisdictions/"
+            + JURISDICTION_ID
+            + "/case-types/"
+            + CASE_TYPE_ID
+            + "/cases/"
+            + CASE_ID
+            + "/events";
     }
 }
