+++ src/contractTest/java/uk/gov/hmcts/reform/et/syaapi/consumer/test/document/PdfGenerationServiceConsumerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.et.syaapi.consumer.document;
+package uk.gov.hmcts.reform.et.syaapi.consumer.test.document;
 
 import au.com.dius.pact.consumer.MockServer;
 import au.com.dius.pact.consumer.dsl.PactDslWithProvider;
@@ -6,7 +6,7 @@ import au.com.dius.pact.consumer.junit5.PactConsumerTestExt;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
 import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
-import au.com.dius.pact.core.model.annotations.PactFolder;
+import au.com.dius.pact.core.model.annotations.PactDirectory;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;
@@ -14,15 +14,13 @@ import org.apache.http.client.fluent.Executor;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.HttpClientBuilder;
-import org.junit.After;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.hmcts.reform.et.syaapi.models.ClaimCaseDocument;
-
-import java.io.IOException;
 
 import static org.junit.Assert.assertEquals;
 import static org.springframework.http.HttpMethod.POST;
@@ -31,7 +29,7 @@ import static org.springframework.http.HttpMethod.POST;
 @ExtendWith(SpringExtension.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @PactTestFor(providerName = "etSyaService_pdfgenerationEndpoint", port = "8891")
-@PactFolder("pacts")
+@PactDirectory("pacts")
 class PdfGenerationServiceConsumerTest {
     private static final String SERVICE_AUTHORIZATION_HEADER = "ServiceAuthorization";
     ObjectMapper objectMapper;
@@ -39,13 +37,13 @@ class PdfGenerationServiceConsumerTest {
     private static final String SOME_SERVICE_AUTH_TOKEN = "someServiceAuthToken";
 
     @BeforeEach
-    void setUpEachTest() throws InterruptedException, IOException {
+    void setUpEachTest() throws InterruptedException {
         Thread.sleep(2000);
         objectMapper = new ObjectMapper();
     }
 
-    @After
-    void teardown() {
+    @AfterEach
+    public void teardown() {
         Executor.closeIdleConnections();
     }
