+++ src/contractTest/java/uk/gov/hmcts/reform/et/syaapi/consumer/test/idam/JwksConsumerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.et.syaapi.consumer.idam;
+package uk.gov.hmcts.reform.et.syaapi.consumer.test.idam;
 
 import au.com.dius.pact.consumer.MockServer;
 import au.com.dius.pact.consumer.dsl.PactDslJsonBody;
@@ -7,8 +7,9 @@ import au.com.dius.pact.consumer.junit5.PactConsumerTestExt;
 import au.com.dius.pact.consumer.junit5.PactTestFor;
 import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
-import au.com.dius.pact.core.model.annotations.PactFolder;
+import au.com.dius.pact.core.model.annotations.PactDirectory;
 import io.restassured.RestAssured;
+import org.assertj.core.api.Assertions;
 import org.json.JSONObject;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -18,25 +19,24 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.util.Map;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.http.HttpMethod.GET;
 import static org.springframework.http.HttpStatus.OK;
 
 @ExtendWith(PactConsumerTestExt.class)
 @ExtendWith(SpringExtension.class)
-@PactFolder("pacts")
-class S2SConsumerTest {
-    private static final String S2S_URL = "/lease";
+@PactDirectory("pacts")
+class JwksConsumerTest {
+    private static final String JWKS_AUTH_URL = "/o/jwks";
 
-    @Pact(provider = "s2-auth-api", consumer = "et_sya_api_service")
-    RequestResponsePact executeServiceAuthApiGetToken(PactDslWithProvider builder) {
+    @Pact(provider = "idam_jwks_api", consumer = "et_sya_api_service")
+    RequestResponsePact executeServiceAuthApiGetToke(PactDslWithProvider builder) {
 
-        Map<String, String> responseHeaders = Map.of(HttpHeaders.AUTHORIZATION, "someToken");
+        Map<String, String> responseHeaders = Map.of(HttpHeaders.AUTHORIZATION, "Bearer UserAuthToken");
 
         return builder
-            .given("a secret exists")
+            .given("a case exists")
             .uponReceiving("Provider receives a token request request from et-sya-api API")
-            .path(S2S_URL)
+            .path(JWKS_AUTH_URL)
             .method(GET.toString())
             .willRespondWith()
             .status(OK.value())
@@ -46,16 +46,15 @@ class S2SConsumerTest {
     }
 
     @Test
-    @PactTestFor(pactMethod = "executeServiceAuthApiGetToken")
+    @PactTestFor(pactMethod = "executeServiceAuthApiGetToke")
     void shouldReceiveTokenAnd200(MockServer mockServer) {
 
         String responseBody = RestAssured
             .given()
             .contentType(MediaType.APPLICATION_FORM_URLENCODED_VALUE)
-            .body(createRequestBody())
             .log().all(true)
             .when()
-            .get(mockServer.getUrl() + S2S_URL)
+            .get(mockServer.getUrl() + JWKS_AUTH_URL)
             .then()
             .statusCode(200)
             .and()
@@ -63,20 +62,17 @@ class S2SConsumerTest {
             .asString();
 
         JSONObject response = new JSONObject(responseBody);
-        assertThat(response).isNotNull();
-        assertThat(response.getString("token")).isNotBlank();
-    }
+        Assertions.assertThat(response).isNotNull();
 
+    }
 
     private PactDslJsonBody createAuthResponse() {
         return new PactDslJsonBody()
-            .stringType("token","someMicroServiceToken");
+            .eachLike("keys")
+                .stringType("kid", "KeyId1")
+                .stringType("kty", "RSA")
+                .stringType("e", "AQAB")
+                .stringType("use", "Public Key Use1")
+                .stringType("n", "someToken");
     }
-
-    private static String createRequestBody() {
-        return new StringBuffer().append("{\"microservice\": \"microServiceName\"))")
-            .append(" \"oneTimePassword\": \"987651\",")
-            .append(" }").toString();
-    }
-
 }
